+++ everyswim/Presentation/Main/Dashboard/Challange/Subviews/AnimateRingUIView.swift
