+++ everyswim/Presentation/Main/Dashboard/RecommandCollection/Model/CommunityCollectionData.swift
@@ -1,0 +1,15 @@
+//
+//  CommunityCollectionData.swift
+//  everyswim
+//
+//  Created by HeonJin Ha on 11/7/23.
+//
+
+import Foundation
+
+struct CommunityCollectionData {
+    let title: String
+    let description: String
+    let url: String
+    let imageUrl: String
+}
