+++ everyswim/Presentation/Main/Dashboard/RecommandCollection/Model/VideoCollectionData.swift
@@ -1,0 +1,15 @@
+//
+//  VideoCollectionData.swift
+//  swimmers
+//
+//  Created by HeonJin Ha on 9/9/23.
+//
+
+import Foundation
+
+struct VideoCollectionData {
+    let id: String
+    let type: String
+    let mediaUrl: String
+    let imageUrl: String
+}
