--- everyswim/Presentation/Main/Dashboard/Subviews/MediaCollection/MediaCollectionData.swift
@@ -1,22 +1,0 @@
-//
-//  MediaCollectionData.swift
-//  swimmers
-//
-//  Created by HeonJin Ha on 9/9/23.
-//
-
-import Foundation
-
-struct MediaCollectionData {
-    let id: String
-    let type: String
-    let mediaUrl: String
-    let imageUrl: String
-}
-
-struct CommunityCollectionData {
-    let title: String
-    let description: String
-    let url: String
-    let imageUrl: String
-}
