--- everyswim/Presentation/Main/Dashboard/Types/WeekdayType.swift
@@ -1,18 +1,0 @@
-//
-//  WeekdayType.swift
-//  swimmers
-//
-//  Created by HeonJin Ha on 2023/06/25.
-//
-
-import Foundation
-
-enum WeekdayType: String {
-    case sun = "일"
-    case mon = "월"
-    case tue = "화"
-    case wed = "수"
-    case thu = "목"
-    case fri = "금"
-    case sat = "토"
-}
